Correct asynchronisity
I’m not sure how it was working before (maybe just by luck?) but with
the prior code, there was no guarantee that the teams request would
finish before the events request did, which meant that the events
could’ve tried to be added before the `teamIdsToRoleNames` map was
ready, which would’ve broke the feed. Also, the server was starting
before the feeds were ready, which meant that requests immediately
after the server started might’ve resulted in invalid responses.