Only add location when it's present
Before, if there was no venue, we we're adding a location as an empty string; now, we leave location undefined on the result unless a venue with an address exists. 

I'm not sure if this actually makes a difference in the event output, or if the `ical` library is smart enough to not output a location field at all when the location is set as an empty string. But, either way, this seems more logically correct (and more robust against changes in the library's behavior).